requirements.txt
paramiko==3.4.0
watchdog==3.0.0
boto3==1.34.34
python-dotenv==1.0.1
schedule==1.2.1 